origin 설정을 Relay 전송에서 분리한다
