+++ apps/app/src/auth/logout.ts
@@ -1,9 +1,9 @@
-import { getWebOrigin } from '@/relay/network';
+import { getPublicWebOrigin } from '@/config/origin';
 
 export const LOGOUT_FAILURE_MESSAGE = '로그아웃하지 못했습니다. 다시 시도해주세요.';
 
 export async function requestWebLogout(): Promise<void> {
-  const response = await fetch(`${getWebOrigin()}/logout`, {
+  const response = await fetch(`${getPublicWebOrigin()}/logout`, {
     cache: 'no-store',
     credentials: 'include',
     method: 'POST',
