+++ apps/app/src/components/post/PostMoreMenu.tsx
@@ -2,7 +2,7 @@ import { Link2 } from 'lucide-react-native';
 import { useCallback } from 'react';
 import { setStringAsync } from '@/components/post/postClipboard';
 import { useToast } from '@/components/ui/ToastProvider';
-import { getWebOrigin } from '@/relay/network';
+import { getPublicWebOrigin } from '@/config/origin';
 import { createPostShareReference } from './postShareReference';
 import type { ActionMenuItem } from '@/components/ui/ActionMenu';
 
@@ -17,7 +17,7 @@ export function usePostMoreMenuItem({ postId, relativeHandle }: Props): ActionMe
   const { showToast } = useToast();
   const copyReference = useCallback(async () => {
     try {
-      const reference = createPostShareReference(getWebOrigin(), relativeHandle, postId);
+      const reference = createPostShareReference(getPublicWebOrigin(), relativeHandle, postId);
       const copied = await setStringAsync(reference);
       if (!copied) {
         throw new Error('Clipboard did not confirm the copy.');
