+++ apps/app/src/relay/network.test.ts
@@ -1,13 +1,6 @@
 import assert from 'node:assert/strict';
 import { describe, it } from 'node:test';
-import {
-  executeGraphQLRequest,
-  formatGraphQLError,
-  getApiOrigin,
-  getWebOrigin,
-  normalizeApiOrigin,
-  normalizeWebOrigin,
-} from './network';
+import { executeGraphQLRequest, formatGraphQLError } from './network';
 
 process.env.EXPO_PUBLIC_API_ORIGIN = 'http://127.0.0.1:4200';
 process.env.EXPO_PUBLIC_WEB_ORIGIN = 'http://127.0.0.1:5173';
@@ -77,25 +70,6 @@ describe('Relay 네트워크', () => {
   });
 });
 
-describe('네이티브 API origin', () => {
-  it('설정된 API origin만 사용한다', () => {
-    assert.equal(getApiOrigin(), 'http://127.0.0.1:4200');
-  });
-
-  it('HTTPS 또는 loopback origin을 정규화한다', () => {
-    assert.equal(
-      normalizeApiOrigin('https://api.kosmo.example/', false),
-      'https://api.kosmo.example',
-    );
-    assert.equal(normalizeApiOrigin('http://127.0.0.1:4200', false), 'http://127.0.0.1:4200');
-  });
-
-  it('path가 있거나 안전하지 않은 remote origin을 기본으로 거부한다', () => {
-    assert.throws(() => normalizeApiOrigin('https://api.kosmo.example/graphql', false));
-    assert.throws(() => normalizeApiOrigin('http://api.kosmo.example', false));
-  });
-});
-
 function stubNavigatorProduct(product: string): () => void {
   const descriptor = Object.getOwnPropertyDescriptor(globalThis, 'navigator');
   Object.defineProperty(globalThis, 'navigator', {
@@ -111,69 +85,3 @@ function stubNavigatorProduct(product: string): () => void {
     }
   };
 }
-
-describe('Web origin', () => {
-  it('build-time 설정 대신 현재 browser origin을 사용한다', () => {
-    const existingWindow = globalThis.window;
-    Object.defineProperty(globalThis, 'window', {
-      configurable: true,
-      value: { location: { origin: 'http://localhost:5173' } },
-    });
-
-    try {
-      assert.equal(getWebOrigin(), 'http://localhost:5173');
-    } finally {
-      if (existingWindow) {
-        Object.defineProperty(globalThis, 'window', {
-          configurable: true,
-          value: existingWindow,
-        });
-      } else {
-        Reflect.deleteProperty(globalThis, 'window');
-      }
-    }
-  });
-
-  it('browser origin이 없을 때 설정된 Web origin으로 fallback한다', () => {
-    const existingWindow = Object.getOwnPropertyDescriptor(globalThis, 'window');
-    Object.defineProperty(globalThis, 'window', {
-      configurable: true,
-      value: { location: {} },
-    });
-    const configured = process.env.EXPO_PUBLIC_WEB_ORIGIN;
-    process.env.EXPO_PUBLIC_WEB_ORIGIN = 'https://configured.example/';
-
-    try {
-      assert.equal(getWebOrigin(), 'https://configured.example');
-    } finally {
-      process.env.EXPO_PUBLIC_WEB_ORIGIN = configured;
-      if (existingWindow) {
-        Object.defineProperty(globalThis, 'window', existingWindow);
-      } else {
-        Reflect.deleteProperty(globalThis, 'window');
-      }
-    }
-  });
-
-  it('HTTPS 또는 loopback origin을 정규화한다', () => {
-    assert.equal(normalizeWebOrigin('https://kosmo.example/', false), 'https://kosmo.example');
-    assert.equal(normalizeWebOrigin('http://127.0.0.1:4173', false), 'http://127.0.0.1:4173');
-  });
-
-  it('path가 있거나 안전하지 않은 remote origin을 기본으로 거부한다', () => {
-    assert.throws(() => normalizeWebOrigin('https://kosmo.example/app', false));
-    assert.throws(() => normalizeWebOrigin('http://kosmo.example', false));
-    assert.equal(normalizeWebOrigin('http://192.0.2.1:4173', true), 'http://192.0.2.1:4173');
-  });
-
-  it('browser 밖에서 origin 설정이 없으면 요청을 차단한다', () => {
-    const configured = process.env.EXPO_PUBLIC_WEB_ORIGIN;
-    delete process.env.EXPO_PUBLIC_WEB_ORIGIN;
-
-    try {
-      assert.throws(() => getWebOrigin());
-    } finally {
-      process.env.EXPO_PUBLIC_WEB_ORIGIN = configured;
-    }
-  });
-});
